+++ App.js
@@ -40,7 +40,7 @@ export default class App extends React.Component {
       // Font.loadAsync({
       //   // This is the font that we are using for our tab bar
       //   ...Icon.Ionicons.font,
-      //   // We include SpaceMono because we use it in HomeScreen.js. Feel free
+      //   // We include SpaceMono because we use it in SplashScreen.js. Feel free
       //   // to remove this if you are not using it in your app
       //   "space-mono": require("./assets/fonts/SpaceMono-Regular.ttf")
       // })
