+++ navigation/AppNavigator.js
@@ -1,23 +1,23 @@
 import React from "react";
 import { createStackNavigator } from "react-navigation";
-import HomeScreen from "../screens/HomeScreen";
+import SplashScreen from "../screens/SplashScreen";
 import LoginScreen from "../screens/LoginScreen";
-import LandingScreen from "../screens/LandingScreen";
+import HomeScreen from "../screens/HomeScreen";
 import Nav from "../components/NavBar/Nav";
 import ProfileScreen from "../screens/ProfileScreen";
 import LeaderboardScreen from "../screens/LeaderboardScreen";
 import LeaderboardsScreen from "../screens/LeaderboardsScreen";
-import editProfile from "../components/Profiles/editProfile";
+import EditProfile from "../components/Profiles/EditProfile";
 import Aftab from "../components/Games/Aftab";
 import Eric from "../components/Games/Eric";
 import Anthony from "../components/Games/Anthony";
 
 export default (AppNavigator = createStackNavigator({
-  Home: { screen: HomeScreen },
+  Splash: { screen: SplashScreen },
   Login: { screen: LoginScreen },
-  Landing: { screen: LandingScreen },
+  Home: { screen: HomeScreen },
   NavBar: { screen: Nav },
-  EditProfile: { screen: editProfile },
+  EditProfile: { screen: EditProfile },
   Profile: { screen: ProfileScreen },
   Leaderboard: { screen: LeaderboardScreen },
   Leadboards: { screen: LeaderboardsScreen },
